vial: explicitly define VIAL_PROTOCOL_VERSION as 32-bit

+++ quantum/vial.h
@@ -18,6 +18,6 @@
 
 #include <inttypes.h>
 
-#define VIAL_PROTOCOL_VERSION 0x00000000
+#define VIAL_PROTOCOL_VERSION ((uint32_t)0x00000000)
 
 void vial_handle_cmd(uint8_t *data, uint8_t length);
